Det går att ta bort onödiga uttag eller överföringar

+++ models/transaction.rb
@@ -2,7 +2,7 @@ require 'sequel'
 
 DB.create_table? :transactions do
 	primary_key :id
-	Integer :sum, null: false
+	Float :sum, null: false
 	Integer :category, null: false, default: 0
 	Time :timestamp, null: false, default: 0
 	String :name, null: false, default: 0
